Add sketch compilation tests
This commit introduces tests for the following pages:
- /libraries
- /user/cb_compile_tester

It tests that every single sketch compiles. For the latter, we keep
a log of all successes and failures in the logs directory.
For some reason, it appears that the vast majority of cb_compile_tester's
sketches are failing to verify. Not sure yet whether that's something to do
with the script, or whether those sketches actually have errors.